requirements/base.txt
attrs==20.2.0
PyQt5==5.15.1
boltons==20.2.1
